pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "felvi_games"
version = "0.1.0"
description = "Gamified Hungarian middle-school entrance exam practice (matek + magyar)"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.35.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pdftotext>=2.2.0",  # conda-forge: conda install -c conda-forge pdftotext
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/felvi_games"]
omit = ["*/app.py"]  # Streamlit UI – skip in unit tests

[project.scripts]
felvi = "felvi_games.cli:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
felvi_games = ["py.typed"]
